refactor: RepositoryTest 패키지 변경
